requirements.txt
streamlit==1.32.0
groq==0.4.2
PyPDF2==3.0.1
langchain==0.1.12
langchain-community==0.0.27
langchain-groq==0.0.1
python-dotenv==1.0.1
sentence-transformers==2.5.1
faiss-cpu==1.8.0 